Convert email fixtures to CRLF instead of just LF
After upgrading to mail 2.7.1 we were seeing a spec fail with:
  NoMethodError: undefined method `content_type' for nil:NilClass
  spec/lib/mail_handler/mail_handler_spec.rb:290

When parsing the mail fixture only one mail part instead of the previous
three, was being detected.

This was caused by mail parsing changes upstream [1] [2] which converted
newlines. The upstream conversion is dependent on if the raw email
contains only ASCII characters or not [3]. In the cast of our
`track-response-ms-bounce` email fixture it contains non-ASCII
characters so the conversion doesn't happen.

Have opted to ensure all our fixture emails are converted to CRLF as
this is who we should expect these to be delivered to us [4].

[1] https://github.com/mikel/mail/commit/f8b57f6 Parse emails with \n newlines if they don't contain binary content
[2] https://github.com/mikel/mail/commit/d17be66 restore LF->CRLF conversion for properly encoded non-binary messages
[3] https://github.com/mikel/mail/blob/2.7.1/lib/mail/utilities.rb#L302-L304
[4] https://tools.ietf.org/html/rfc2822#section-2.1